+++ models/category/src/category-search/presets/with-parent.ts
@@ -1,0 +1,8 @@
+import { TCategorySearchBuilder } from '../../types';
+import CategorySearch from '../builder';
+
+const withParent = (): TCategorySearchBuilder => {
+  return CategorySearch().parent(CategorySearch());
+};
+
+export default withParent;