+++ models/platform-limits/src/business-units/presets/index.ts
@@ -1,0 +1,5 @@
+import withLimit from './with-limit';
+
+export default {
+  withLimit,
+};