+++ models/platform-limits/src/general/builder.ts
@@ -1,0 +1,15 @@
+import { Builder } from '@commercetools-test-data/core';
+import generator from './generator';
+import transformers from './transformers';
+import {
+  TProjectCustomLimitsProjection,
+  TCreateProjectCustomLimitsProjectionBuilder,
+} from './types';
+
+const Model: TCreateProjectCustomLimitsProjectionBuilder = () =>
+  Builder<TProjectCustomLimitsProjection>({
+    generator,
+    transformers,
+  });
+
+export default Model;