+++ models/platform-limits/src/shipping-methods/presets/with-limit-and-current.ts
@@ -1,0 +1,7 @@
+import * as LimitWithCurrent from '../../limit-with-current';
+import ShippingMethodsPlatformLimits from '../builder';
+
+const withLimitAndCurrent = () =>
+  ShippingMethodsPlatformLimits().total(LimitWithCurrent.random());
+
+export default withLimitAndCurrent;