+++ models/platform-limits/src/stores/presets/with-limit-and-current.ts
@@ -1,0 +1,11 @@
+import * as Limit from '../../limit';
+import * as LimitWithCurrent from '../../limit-with-current';
+import StoresPlatformLimits from '../builder';
+
+const withLimitAndCurrent = () =>
+  StoresPlatformLimits()
+    .total(LimitWithCurrent.random())
+    .inventorySupplyChannels(Limit.random())
+    .productDistributionChannels(Limit.random());
+
+export default withLimitAndCurrent;